Move explorers to datastore
